Organize BLE DBus objects
