+++ rpi/ble/DBusObjects/Characteristic.py
