+++ rpi/ble/DBusObjects/Descriptor.py
