+++ rpi/ble/DBusObjects/TestAdvertisement.py
@@ -1,0 +1,12 @@
+from .Advertisement import Advertisement
+
+class TestAdvertisement(Advertisement):
+  def __init__(self, bus, index):
+    Advertisement.__init__(self, bus, index, 'peripheral')
+    self.add_service_uuid('180D')
+    self.add_service_uuid('180F')
+    self.add_manufacturer_data(0xffff, [0x00, 0x01, 0x02, 0x03, 0x04])
+    self.add_service_data('9999', [0x00, 0x01, 0x02, 0x03, 0x04])
+    self.add_local_name('YUUUUH BOIOIOI')
+    self.include_tx_power = True
+    self.add_data(0x26, [0x01, 0x01, 0x00])
