+++ rpi/ble/DBusObjects/TestService/CharacteristicUserDescriptionDscr.py
