+++ rpi/ble/DBusObjects/errors.py
